bayesian_optimization/requirements.txt
numpy
scipy
matplotlib
python-mnist
pytest
tqdm
tensorflow
ffmpeg
